Bug 799432 - Backout changeset 30ab4e1d4874 and changeset dbcc29e9fcf6 from bug 777292 since they're no longer necessary with bug 795238
UltraBlame original commit: f9bb37e4c7b679647ae468059d3cd1d3a8b7fd12

+++ js/xpconnect/src/XPCComponents.cpp
@@ -1404,10 +1404,7 @@ nsXPCComponents_Results::NewResolve(nsIXPConnectWrappedNative *wrapper,
         nsresult rv;
         while (nsXPCException::IterateNSResults(&rv, &rv_name, nullptr, &iter)) {
             if (!strcmp(name.ptr(), rv_name)) {
-                
-                
-                
-                jsval val = JS_NumberValue((double)(uint32_t)rv);
+                jsval val = JS_NumberValue((double)rv);
 
                 *objp = obj;
                 if (!JS_DefinePropertyById(cx, obj, id, val,
@@ -4734,10 +4731,7 @@ nsXPCComponents::GetProperty(nsIXPConnectWrappedNative *wrapper,
 
     nsresult rv = NS_OK;
     if (doResult) {
-        
-        
-        
-        *vp = JS_NumberValue((double)(uint32_t) res);
+        *vp = JS_NumberValue((double) res);
         rv = NS_SUCCESS_I_DID_SOMETHING;
     }
